backend/requirements.txt
asgiref
awesomeversion
beautifulsoup4
cryptography
dj-database-url
Django==3.1.7
django-cors-headers
django-extensions
django-js-asset
django-filter
django-rest-auth[with_social]
djangorestframework
gunicorn
psycopg2-binary
pytz
requests
six
whitenoise
pyimgur
